ijunavi/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
pypdf
tiktoken
gradio
python-dotenv
pandas
#zipfile
requests